requirements.txt
autogen>=0.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0 